requirements.txt
certifi==2018.11.29
chardet==3.0.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django==2.1.4
idna==2.8
pytz==2018.7
requests==2.21.0
sqlparse==0.2.4
urllib3==1.24.1